Variable axis properties (#99)
* Bumped to next SNAPTHOT version

* Added Axis type and refactored how IDs are made

* Added VariableConnection validation on the Axis property

* Added VE_VariableGroup_4 to verify that a VariableGroup does not contain variables with identical axis properties

* Added VE_VariableGroupConnection_4 which checks a VariableGroupConnection for variables containing incompatible axis properties

* Renamed osp_labelled_modified.ttl to osp.ttl

* Reran JAXB model generator and updated README.md with how to run this command

* Update dependency on FMI4j

* Added support for VE_VariableGroup4 ontology implementation

* Ontology debugging and addition of VE_Variable classes.

* Removed ontology debugging functionality

* Added support for VE_VariableGroupConnection_4 ontolgy implementation + update to onotology.  Currently using Java implmentation. Ontology requires further work before ontology implementation can be used.

* Updated readme to reflect changes in version 1.0.1. Updated name of xsd from 1.0.0 to 1.0.1. https://open-simulation-platform.com/OspModelDescription/1.0.0 must be updated to 1.0.1.

* Change to OspModelDescription/1.0.1 in OspModelDescription.xml files. Rerun xjc (com.opensimulationplatform.modeldescription classes).

* Prepend project in properties to fix warnings

* Update link to xsd in README.md

* SNAPTHOT -> SNAPSHOT

* fmu-proxy -> proxyfmu. Handle URI decoding in DefaultFmuLocator. No support for remote FMUs yet.

* Using OspSystemStructure.xsd from libcosim version 0.10.0

+++ osp-validator-core/src/test/java/com/opensimulationplatform/core/validation/variableconnection/VE_VariableConnection_2_Test.java
@@ -44,7 +44,7 @@ public class VE_VariableConnection_2_Test {
 
     Variable vb = new Variable();
     vb.setName("vb");
-    vb.setType(Variable.Type.INTEGER);
+    vb.setType(Variable.Type.REAL);
     vb.setCausality(Variable.Causality.INPUT);
 
     VariableConnection vc = new VariableConnection();
